[#40-1] feat : mysql, redis, common 모듈 생성 및 클래스 분리
