+++ build.gradle
@@ -20,6 +20,7 @@ allprojects {
 // 서브 프로젝트 전체에 적용.
 subprojects {
 	apply plugin: 'java'
+	apply plugin: 'java-library'
 	apply plugin: 'org.springframework.boot'
 	apply plugin: 'io.spring.dependency-management'
 
@@ -34,9 +35,6 @@ subprojects {
 	}
 
 	dependencies {
-		implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-		implementation 'org.redisson:redisson-spring-boot-starter:3.21.1'
-		implementation 'org.springframework.boot:spring-boot-starter-data-redis'
 		implementation 'org.springframework.boot:spring-boot-starter-validation'
 		implementation 'org.springframework.cloud:spring-cloud-starter-vault-config:3.1.1'
 
@@ -57,8 +55,6 @@ subprojects {
 				'org.projectlombok:lombok-mapstruct-binding:0.2.0'
 		)
 
-		runtimeOnly 'com.h2database:h2'
-		runtimeOnly 'com.mysql:mysql-connector-j'
 		testImplementation 'org.springframework.boot:spring-boot-starter-test'
 		testImplementation 'org.mockito:mockito-core:3.11.2'
 		testImplementation 'org.junit.jupiter:junit-jupiter-api:5.9.2'
@@ -82,5 +78,19 @@ project(':goodchoice-api') {
 		implementation project(':goodchoice-order')
 		implementation project(':goodchoice-coupon')
 		implementation project(':goodchoice-item')
+		implementation project(':goodchoice-redis')
+		implementation project(':goodchoice-mysql')
 	}
 }
+
+//project(':goodchoice-member') {
+//	dependencies {
+//		implementation project(':goodchoice-mysql')
+//	}
+//}
+//
+//project(':goodchoice-member') {
+//	dependencies {
+//		implementation project(':goodchoice-mysql')
+//	}
+//}
