+++ goodchoice-api/src/main/java/com/flab/goodchoiceapi/common/AuthorizationFilter.java
@@ -16,7 +16,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
-@Component
+//@Component
 @RequiredArgsConstructor
 public class AuthorizationFilter implements Filter {
