--- goodchoice-api/src/main/java/com/flab/goodchoiceapi/common/db/DynamicRoutingDataSource.java
@@ -1,14 +1,0 @@
-package com.flab.goodchoiceapi.common.db;
-
-import org.springframework.jdbc.datasource.lookup.AbstractRoutingDataSource;
-
-import static org.springframework.transaction.support.TransactionSynchronizationManager.isCurrentTransactionReadOnly;
-
-public class DynamicRoutingDataSource extends AbstractRoutingDataSource {
-
-    @Override
-    protected Object determineCurrentLookupKey() {
-        return isCurrentTransactionReadOnly() ? "slave" : "master";
-    }
-
-}
