--- goodchoice-api/src/main/java/com/flab/goodchoiceapi/global/config/aop/RedissonConfig.java
@@ -1,30 +1,0 @@
-package com.flab.goodchoiceapi.global.config.aop;
-
-import org.redisson.Redisson;
-import org.redisson.api.RedissonClient;
-import org.redisson.config.Config;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class RedissonConfig {
-
-    @Value("${spring.redis.host}")
-    private String redisHost;
-
-    @Value("${spring.redis.port}")
-    private int redisPort;
-
-    private static final String REDISSON_HOST_PREFIX = "redis://";
-
-    @Bean
-    public RedissonClient redissonClient() {
-        RedissonClient redisson = null;
-        Config config = new Config();
-        config.useSingleServer().setAddress(REDISSON_HOST_PREFIX + redisHost + ":" + redisPort);
-        redisson = Redisson.create(config);
-        return redisson;
-    }
-
-}
