--- goodchoice-api/src/main/java/com/flab/goodchoiceapi/global/config/aop/RedissonLockAop.java
@@ -1,71 +1,0 @@
-package com.flab.goodchoiceapi.global.config.aop;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.aspectj.lang.ProceedingJoinPoint;
-import org.aspectj.lang.annotation.Around;
-import org.aspectj.lang.annotation.Aspect;
-import org.aspectj.lang.reflect.MethodSignature;
-import org.redisson.api.RLock;
-import org.redisson.api.RedissonClient;
-import org.springframework.stereotype.Component;
-
-import java.lang.reflect.Method;
-
-@Slf4j
-@Aspect
-@Component
-@RequiredArgsConstructor
-public class RedissonLockAop {
-
-    private final RedissonClient redissonClient;
-
-    private final RedissonCallTransaction redissonCallTransaction;
-
-    @Around("@annotation(com.flab.goodchoiceapi.global.config.aop.DistributedLock)")
-    public Object lock(final ProceedingJoinPoint joinPoint) throws Throwable {
-        MethodSignature signature = (MethodSignature) joinPoint.getSignature();
-        Method method = signature.getMethod();
-        DistributedLock distributedLock = method.getAnnotation(DistributedLock.class);
-
-        String key = this.createKey(signature.getParameterNames(), joinPoint.getArgs(), distributedLock.key());
-        RLock rLock = redissonClient.getLock(key);
-        try {
-            boolean isPossible = rLock.tryLock(distributedLock.waitTime(), distributedLock.leaseTime(), distributedLock.timeUnit());
-            if (!isPossible) {
-                return false;
-            }
-
-            log.info("Redisson Lock Key : {}", key);
-            return redissonCallTransaction.proceed(joinPoint);
-        } catch (InterruptedException e) {
-            throw new InterruptedException();
-        } finally {
-            try {
-                rLock.unlock();   // (4)
-            } catch (IllegalMonitorStateException e) {
-                log.info("Redisson Lock Already UnLock {} {}");
-            }
-        }
-
-    }
-
-    /**
-     * Redisson Key Create
-     * @param parameterNames
-     * @param args
-     * @param key
-     * @return
-     */
-    private String createKey(String[] parameterNames, Object[] args, String key) {
-        String resultKey = key;
-        for (int i = 0; i < parameterNames.length; i++) {
-            if (parameterNames[i].equals(key)) {
-                resultKey += args[i];
-                break;
-            }
-        }
-        return resultKey;
-    }
-
-}
