+++ goodchoice-api/src/main/resources/application.yml
@@ -3,6 +3,9 @@ server:
 spring:
   profiles:
     active: local
+#    include:
+#      - mysql
+#      - redis
   config:
     #   activate:
     #     on-profile: loc
