+++ goodchoice-common/build.gradle
@@ -1,0 +1,3 @@
+dependencies {
+}
+tasks.register("prepareKotlinBuildScriptModel") {}
