--- goodchoice-coupon/src/main/java/com/flab/goodchoicecoupon/infrastructure/repositories/AppliedUserRepository.java
@@ -1,6 +1,0 @@
-package com.flab.goodchoicecoupon.infrastructure.repositories;
-
-public interface AppliedUserRepository {
-
-    Long addRedisSet(String key, String memberId);
-}
