--- goodchoice-coupon/src/testFixtures/java/com/flab/goodchoicecoupon/infrastructure/FakeAppliedUserRepository.java
@@ -1,10 +1,0 @@
-package com.flab.goodchoicecoupon.infrastructure;
-
-import com.flab.goodchoicecoupon.infrastructure.repositories.AppliedUserRepository;
-
-public class FakeAppliedUserRepository implements AppliedUserRepository {
-    @Override
-    public Long addRedisSet(String key, String memberId) {
-        return null;
-    }
-}
