+++ goodchoice-mysql/build.gradle
@@ -1,0 +1,15 @@
+dependencies {
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+
+    runtimeOnly 'com.h2database:h2'
+    runtimeOnly 'com.mysql:mysql-connector-j'
+}
+tasks.register("prepareKotlinBuildScriptModel") {}
+
+bootJar {
+    enabled = false
+}
+jar {
+    enabled = true
+}
