+++ goodchoice-mysql/src/main/java/com/flab/goodchoicemysql/common/db/DynamicRoutingDataSource.java
@@ -1,0 +1,14 @@
+package com.flab.goodchoicemysql.common.db;
+
+import org.springframework.jdbc.datasource.lookup.AbstractRoutingDataSource;
+
+import static org.springframework.transaction.support.TransactionSynchronizationManager.isCurrentTransactionReadOnly;
+
+public class DynamicRoutingDataSource extends AbstractRoutingDataSource {
+
+    @Override
+    protected Object determineCurrentLookupKey() {
+        return isCurrentTransactionReadOnly() ? "slave" : "master";
+    }
+
+}
