+++ goodchoice-mysql/src/main/resources/application-mysql.yml
@@ -1,0 +1,31 @@
+server:
+  forward-headers-strategy: framework
+
+spring:
+  config:
+    activate:
+      on-profile: mysql
+
+  datasource:
+    hikari:
+      master:
+        jdbc-url: jdbc:mysql://118.67.135.217:32770/goodchoice
+        username: root
+        password: 1234
+        driver-class-name: com.mysql.cj.jdbc.Driver
+      slave:
+        jdbc-url: jdbc:mysql://118.67.135.217:32769/goodchoice
+        username: root
+        password: 1234
+        driver-class-name: com.mysql.cj.jdbc.Driver
+
+  jpa:
+    defer-datasource-initialization: true
+    show-sql: true
+    database-platform: org.hibernate.dialect.MySQL5Dialect
+    hibernate:
+      #      ddl-auto: create
+      ddl-auto: none
+    properties:
+      hibernate:
+        format_sql: true
