+++ goodchoice-redis/build.gradle
@@ -1,0 +1,12 @@
+plugins {
+    id 'java-test-fixtures'
+    id 'java-library'
+    id 'maven-publish'
+}
+
+dependencies {
+    implementation 'org.springframework.boot:spring-boot-starter-aop'
+    implementation 'org.springframework.boot:spring-boot-starter-data-redis'
+    implementation 'org.redisson:redisson-spring-boot-starter:3.21.1'
+}
+tasks.register("prepareKotlinBuildScriptModel") {}
