+++ goodchoice-redis/src/main/java/com/flab/goodchoiceredis/common/aop/RedissonCallTransaction.java
@@ -1,0 +1,22 @@
+package com.flab.goodchoiceredis.common.aop;
+
+import org.aspectj.lang.ProceedingJoinPoint;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
+
+@Component
+public class RedissonCallTransaction {
+
+    /**
+     * 부모트랜잭션의 유무와 관계없이 동시성에 대한 처리는 별도의 트랜잭션으로 동작하기 위함
+     * @param joinPoint
+     * @return
+     * @throws Throwable
+     */
+    @Transactional(propagation = Propagation.REQUIRES_NEW)
+    public Object proceed(final ProceedingJoinPoint joinPoint) throws Throwable {
+        return joinPoint.proceed();
+    }
+
+}
