+++ goodchoice-redis/src/main/java/com/flab/goodchoiceredis/infrastructure/AppliedUserRepository.java
@@ -1,0 +1,6 @@
+package com.flab.goodchoiceredis.infrastructure;
+
+public interface AppliedUserRepository {
+
+    Long addRedisSet(String key, String memberId);
+}
