+++ goodchoice-redis/src/testFixtures/java/com/flab/goodchoiceredis/infrastructure/FakeAppliedUserRepository.java
@@ -1,0 +1,8 @@
+package com.flab.goodchoiceredis.infrastructure;
+
+public class FakeAppliedUserRepository implements AppliedUserRepository {
+    @Override
+    public Long addRedisSet(String key, String memberId) {
+        return null;
+    }
+}
