+++ settings.gradle
@@ -13,3 +13,6 @@ include 'goodchoice-coupon'
 include 'goodchoice-order'
 include 'goodchoice-item'
 include 'goodchoice-batch'
+include 'goodchoice-redis'
+include 'goodchoice-mysql'
+include 'goodchoice-common'
